+++ lib/src/models/yaml_comment.dart
@@ -1,62 +1,82 @@
+import 'dart:math' show Random;
+
 /// Represents a YAML comment.
 class YamlComment {
-  static String key = 'COMMENT';
   final String text;
-  final int linesMaxlength;
   final int indentLevel;
+  final String? lineAfter;
+  final int lineKeyExistsCount;
+
+  String get key =>
+      '__comment_${Random().nextInt(100)}_${DateTime.now().millisecondsSinceEpoch}';
 
   /// Creates a new instance of [YamlComment].
   ///
   /// The [text] parameter specifies the comment text.
-  /// The [linesMaxlength] parameter defines the maximum length of each line when splitting the comment into multiple lines.
   /// The [indentLevel] parameter specifies the level of indentation for the comment.
+  /// **Ignore this property**: The [lineAfter] property is used by the YamlMagic package to calculate and define the position of this comment.
+  /// **Ignore this property**: The [lineKeyExistsCount] property is used by the YamlMagic package to calculate and define the position of this comment.
+  ///
   const YamlComment(
     this.text, {
-    this.linesMaxlength = 0,
     this.indentLevel = 0,
+    this.lineAfter,
+    this.lineKeyExistsCount = 1,
   });
 
   @override
   String toString() {
     final indent = '  ' * indentLevel;
-    final lines = <String>[];
-
-    if (linesMaxlength > 0) {
-      final words = text.split(' ');
-      final buffer = StringBuffer();
-
-      for (final word in words) {
-        if ((buffer.length + word.length + 1) <= linesMaxlength) {
-          buffer.write('$word ');
-        } else {
-          lines.add(buffer.toString().trim());
-          buffer.clear();
-          buffer.write('$indent $word ');
-        }
-      }
-
-      if (buffer.isNotEmpty) {
-        lines.add(buffer.toString().trim());
-      }
-    } else {
-      lines.add(text);
-    }
-
-    return lines.map((line) => '$indent# $line').join('\n');
+    return text
+        .split('\n')
+        .map((line) => '$indent# $line')
+        .join('\n')
+        .trimRight();
   }
+
+  Map<String, YamlComment> toMap() => {key: this};
 
   /// Creates a new [YamlComment] instance with
   /// the specified properties updated.
   YamlComment copyWith({
     String? text,
-    int? linesMaxlength,
     int? indentLevel,
+    String? lineAfter,
+    int? lineKeyExistsCount,
   }) {
     return YamlComment(
       text ?? this.text,
-      linesMaxlength: linesMaxlength ?? this.linesMaxlength,
       indentLevel: indentLevel ?? this.indentLevel,
+      lineAfter: lineAfter ?? this.lineAfter,
+      lineKeyExistsCount: lineKeyExistsCount ?? this.lineKeyExistsCount,
     );
+  }
+
+  String? get lineKey => lineAfter == null ? null : getKeyFromLine(lineAfter!);
+
+  /// Retrieve a yaml key string from [line].
+  static String? getKeyFromLine(String line) {
+    final trimmedLine = line.trim();
+    final indexOfColon = trimmedLine.indexOf(':');
+    if (indexOfColon != -1) {
+      String lineSubstring = trimmedLine.substring(0, indexOfColon).trim();
+      if (lineSubstring.startsWith('- ')) {
+        lineSubstring = lineSubstring.substring(2);
+      }
+      return lineSubstring;
+    }
+    return null;
+  }
+
+  String? get lineValue {
+    if (lineAfter != null) {
+      final trimmedLine = lineAfter!.trim();
+      final indexOfColon = trimmedLine.indexOf(':');
+      if (indexOfColon != -1) {
+        return trimmedLine.substring(indexOfColon + 1).trim();
+      }
+    }
+    return null;
   }
 
   @override
@@ -65,11 +85,15 @@ class YamlComment {
 
     return other is YamlComment &&
         other.text == text &&
-        other.linesMaxlength == linesMaxlength &&
-        other.indentLevel == indentLevel;
+        other.indentLevel == indentLevel &&
+        other.lineAfter == lineAfter &&
+        other.lineKeyExistsCount == lineKeyExistsCount;
   }
 
   @override
   int get hashCode =>
-      text.hashCode ^ linesMaxlength.hashCode ^ indentLevel.hashCode;
+      text.hashCode ^
+      indentLevel.hashCode ^
+      lineAfter.hashCode ^
+      lineKeyExistsCount.hashCode;
 }